Refactor: Rename `gameEnd` package to `gameResult`
This commit renames the package `com.amsterdam.viewmodel.gameEnd` to `com.amsterdam.viewmodel.gameResult` and updates all relevant imports.

**Key changes:**

- Renamed `ResultViewModel` to `GameResultViewModel`.
- Renamed `StatCard.kt` and moved it from `ui/src/main/java/com/amsterdam/ui/screens/letsPlay/component` to `ui/src/main/java/com/amsterdam/ui/screens/gameResult/component`.
- Updated import paths for `GameResultArgs`, `ResultInteractionListener`, `ResultScreenData`, `ResultSideEffect`, and `ResultUiState` across multiple files.
- Updated `NotEnoughPointsDialog` import path in various game screens.
- In `app/src/main/java/com/amsterdam/aflami/di/viewModelModule.kt`:
    - Updated `GameResultArgs` import.
    - Updated `GameGenreArgs` import from `com.amsterdam.viewmodel.game.whichGenre` to `com.amsterdam.viewmodel.guessWhichGenre`.
- In `ui/src/main/java/com/amsterdam/ui/components/movieAndTvShowDetails/ReviewCard.kt`:
    - Removed unused imports for `AflamiTheme`, `ThemeAndLocalePreviews`, and `SeriesDetailsUiState.ReviewTvShowUiState`.
