+++ viewModel/src/main/java/com/amsterdam/viewmodel/guessCharacterGame/GuessCharacterGameEffect.kt
@@ -1,6 +1,6 @@
 package com.amsterdam.viewmodel.guessCharacterGame
 
-import com.amsterdam.viewmodel.gameEnd.ResultScreenData
+import com.amsterdam.viewmodel.gameResult.ResultScreenData
 
 
 interface GuessCharacterGameEffect {
