+++ viewModel/src/main/java/com/amsterdam/viewmodel/guessReleseDateGame/GuessReleaseYearGameEffect.kt
@@ -1,6 +1,6 @@
 package com.amsterdam.viewmodel.guessReleseDateGame
 
-import com.amsterdam.viewmodel.gameEnd.ResultScreenData
+import com.amsterdam.viewmodel.gameResult.ResultScreenData
 
 
 interface GuessReleaseYearGameEffect {
