+++ viewModel/src/main/java/com/amsterdam/viewmodel/guessWhichGenre/GenreGameEffect.kt
@@ -1,6 +1,6 @@
-package com.amsterdam.viewmodel.game.whichGenre
+package com.amsterdam.viewmodel.guessWhichGenre
 
-import com.amsterdam.viewmodel.gameEnd.ResultScreenData
+import com.amsterdam.viewmodel.gameResult.ResultScreenData
 
 sealed interface GenreGameEffect {
     data object CancelGame : GenreGameEffect
