+++ viewModel/src/main/java/com/amsterdam/viewmodel/guessWhichGenre/GenreGameUiState.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.viewmodel.game.whichGenre
+package com.amsterdam.viewmodel.guessWhichGenre
 
 import com.amsterdam.domain.exceptions.AflamiException
 import com.amsterdam.entity.category.MovieGenre
